refactor: split decision.rs -- extract build_actions.rs to keep all ai_player submodules <1000 lines
